+++ lizmap/lizmap_api/config.py
@@ -172,9 +172,11 @@ class LizmapConfig:
             'draw': {
                 'wType': 'checkbox', 'type': 'boolean', 'default': False
             },
-            # 'print': {
-            #     'wType': 'checkbox', 'type': 'boolean', 'default': False
-            # },
+            # Deprecated since LWC 3.7.0
+            # There is a new "print" panel
+            'print': {
+                'wType': 'checkbox', 'type': 'boolean', 'default': False
+            },
             'measure': {
                 'wType': 'checkbox', 'type': 'boolean', 'default': False
             },