+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * Group wizard
 * Fix the converter from Lizmap HTML popup to QGIS HTML popup when the alias/field has an underscore or other 
   accented characters.
+* Add the Lizmap Web Client target version in the copy/paste text available on a server
 
 ## 3.9.10 - 2023-01-05